Finished tests

+++ index.js
@@ -52,17 +52,20 @@ app.get(`/users/:id`, async (req, res) => {
 app.post(`/users`, async (req, res) => {
     try{
         const {Name, Age, Email} = req.body;
+        
+        console.log(req.body);
+
         if(!Name || !Age || !Email){
             console.log("Variables not filled")
-            return
+            return res.status(400).json({error: "Missing required fields"}); //Sier her hvilken error jeg forventer
         }
         let sqlInput = "INSERT INTO Users (Name, Age, Email) VALUES (@Name, @Age, @Email)";
         const pool=await sql.connect(config);
-        let data = await pool.request().input("Name", sql.NVarChar, Name). input("Age", sql.Int, Age).input("Email", sql.NVarChar, Email).query(sqlInput);
+        await pool.request().input("Name", sql.NVarChar, Name). input("Age", sql.Int, Age).input("Email", sql.NVarChar, Email).query(sqlInput);
         
         console.log("new user added")
 
-        return res.json(data.recordset)
+        return res.status(201).json({Name, Age, Email})
     }
     catch(err){
         console.log("Error adding user", err)
@@ -79,6 +82,7 @@ app.put(`/users/:id`, async (req, res) => {
         await pool.request().input("id", sql.Int, req.params.id).input("Name", sql.NVarChar, Name).input("Age", sql.Int, Age).input("Email", sql.NVarChar, Email).query(sqlInput);
         
         console.log(`User with id: ${userId} has been updated`)
+        return res.json({Name, Age, Email})
     }
     catch{
         console.error("Error editing user");
@@ -88,15 +92,24 @@ app.put(`/users/:id`, async (req, res) => {
 app.delete(`/users/:id`, async (req, res) => {
     try{
         let userId = parseInt(req.params.id);
+        if(isNaN(userId)){
+            return res.status(400).json({error: "Invalid user ID"});
+        };
         let sqlInput = `DELETE FROM Users WHERE ID = ${userId}`;
 
         const pool = await sql.connect(config);
-        await pool.request().input("id", sql.Int, req.params.id).query(sqlInput);
+        const result = await pool.request().input("id", sql.Int, req.params.id).query(sqlInput);
 
-        console.log(`User with id: ${userId} has been deleted`)
+        if(result.rowsAffected[0] === 0){
+            return res.status(404).json({error: "User not found"});
+        }
+
+        console.log(`User with id: ${userId} has been deleted`);
+        return res.status(200).json({message: `User with id: ${userId} has been deleted`});
     }
     catch{
-        console.log(`Error deleting user with id: ${userId}`)
+        console.log(`Error deleting user with id: ${userId}`);
+        return res.status(500).json();
     }
 })
